backend/voice_call_service/pyproject.toml
[project]
name = "voice-call-service"
version = "0.1.0"
description = "Minimal FastAPI backend for canvas-scoped LiveKit voice rooms"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.135.3",
    "livekit-agents[openai]~=1.4",
    "livekit-api>=1.1.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.43.0",
]

[project.scripts]
voice-agent-worker = "voice_agent_worker:main"

[dependency-groups]
dev = [
    "basedpyright>=1.39.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
include = [
    "app",
    "tests",
]
exclude = [
    ".venv",
    "__pycache__",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
